separate configs and use inheritance
